Reverted caracal branch to revision 1816.

+++ ocelot/cuda/test/functions/indirectCallOcelot.cpp
@@ -28,8 +28,7 @@ int main(int argc, char *arg[]) {
 	
 	result = cudaMalloc((void **)&A_gpu, bytes);
 	if (result != cudaSuccess) {
-		printf("cudaMalloc() - failed to allocate %d on the device \n",
-			(int)bytes);
+		printf("cudaMalloc() - failed to allocate %d on the device \n", bytes);
 		return 2;
 	}
 	
@@ -40,8 +39,7 @@ int main(int argc, char *arg[]) {
 	
 	result = cudaMemcpy(A_gpu, A_cpu, bytes, cudaMemcpyHostToDevice);
 	if (result != cudaSuccess) {
-		printf("cudaMemcpy() - failed to copy %d bytes TO the device\n",
-			(int)bytes);
+		printf("cudaMemcpy() - failed to copy %d bytes TO the device\n", bytes);
 		return 2;
 	}
 	
@@ -63,8 +61,7 @@ int main(int argc, char *arg[]) {
 	
 	result = cudaMemcpy(A_cpu, A_gpu, bytes, cudaMemcpyDeviceToHost);
 	if (result != cudaSuccess) {
-		printf("cudaMemcpy() - failed to copy %d bytes FROM the device\n",
-			(int)bytes);
+		printf("cudaMemcpy() - failed to copy %d bytes FROM the device\n", bytes);
 		return 2;
 	}
